🚚 Move: 그룹 선택 페이지 이동
