+++ src/router.tsx
@@ -8,9 +8,9 @@ import LandingPage from '@/pages/LandingPage';
 import RegisterPage from '@/pages/RegisterPage';
 import SurveyIntroPage from '@/pages/SurveyIntroPage';
 import SurveyPage from '@/pages/SurveyPage';
-import GroupCreatePage from '@/pages/GroupCreatePage';
-import GroupSelectPage from '@/pages/GroupSelectPage';
-import GroupSettingPage from '@/pages/GroupSettingPage';
+import GroupCreatePage from '@/pages/group/GroupCreatePage';
+import GroupSelectPage from '@/pages/group/GroupSelectPage';
+import GroupSettingPage from '@/pages/group/GroupSettingPage';
 import HomePage from '@/pages/home/HomePage';
 import WeeklyStatisticsPage from '@/pages/statistics/WeeklyStatisticsPage';
 import MonthlyStatisticsPage from '@/pages/statistics/MonthlyStatisticsPage';
@@ -21,7 +21,7 @@ import AccountManagePage from '@/pages/AccountManagePage';
 import LeavePage from '@/pages/LeavePage';
 import HouseWorkStepOnePage from '@/pages/housework/HouseWorkStepOnePage';
 import HouseWorkStepTwoPage from '@/pages/housework/HouseWorkStepTwoPage';
-import GroupInviteReceivePage from '@/pages/GroupInviteReceivePage';
+import GroupInviteReceivePage from '@/pages/group/GroupInviteReceivePage';
 import ScrollToTop from '@/components/common/scroll/ScrollToTop';
 import { Toaster } from '@/components/common/ui/toaster';
 import PrivateRouter from '@/components/PrivateRouter';
